backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2

# WhisperX (faster-whisper + speaker diarization via pyannote-audio)
whisperx>=3.1.1

# Speaker diarization
pyannote.audio>=3.1

# Audio processing
pydub==0.25.1
ffmpeg-python==0.2.0

# LLM APIs
openai==1.51.0
anthropic==0.36.0
google-generativeai==0.8.2

# Google Drive
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
jinja2==3.1.4
httpx==0.27.2
